passed bash script

+++ project_1/command.c
@@ -14,11 +14,12 @@ void listDir() {
     struct dirent *entry;
 
     while ((entry = readdir(current)) != NULL) {
-        char name[256];
+        char name[257];
         snprintf(name, sizeof(name), "%s ", entry->d_name);
         write(1,name,strlen(name));
     }
-    write(1, "\n", 1); 
+    write(1, "\n", 1);
+    closedir(current);
 }
 
 
@@ -68,7 +69,7 @@ void copyFile(char *sourcePath, char *destinationPath){
     char* filename = basename(sourcePath);
     int dest;
 
-    if (S_ISDIR(path_stat.st_mode)){
+    if (result == 0 && S_ISDIR(path_stat.st_mode)){
         char dest_path[512];
         snprintf(dest_path, sizeof(dest_path), "%s/%s", destinationPath, basename(sourcePath));
         dest = open(dest_path, O_WRONLY | O_CREAT | O_TRUNC, checker.st_mode & 0777);
@@ -98,31 +99,21 @@ void moveFile(char *sourcePath, char *destinationPath){
         return;
     }
 
-    copyFile(sourcePath,destinationPath);
-    deleteFile(sourcePath);
+    if(sourcePath != destinationPath)
+    {
+        copyFile(sourcePath,destinationPath);
+        deleteFile(sourcePath);
+    }
 }
 
 
 void deleteFile(char *filename){
-    DIR *current = opendir("."); 
-    struct dirent *entry;
-    char* file = basename(filename);
-    int found = 0;
+    int status = remove(filename);
 
-    while ((entry = readdir(current)) != NULL) {
-        if(strcmp(entry->d_name,file) == 0){
-            found = 1;
-            break;
-        }
-    }
-
-    if(found == 0){
+    if(status == -1){
         char error[256];
         snprintf(error, sizeof(error), "rm: cannot remove ‘%s’: No such file or directory\n", filename);
         write(1,error,strlen(error));
-    }
-    else{
-        remove(filename);
     }
 }
 
@@ -143,5 +134,6 @@ void displayFile(char *filename){
     while ((bytes = read(src, buffer, sizeof(buffer))) > 0) {
         write(1, buffer, bytes);
     }
-    write(1, "\n", 1); 
+    
+    close(src);
 }
